fix: rename todolist related files and variables for consistency
